release: 11_00_00_14

+++ mcal/Fls/V2/Fls_Brd_Nor_Ospi.c
@@ -1148,7 +1148,7 @@ uint8 Fls_norBlankCheck(uint32 actualChunkSize)
     uint32 addr         = Fls_DrvObj.flashAddr;
     uint32 len          = actualChunkSize;
     uint8 *readData_buf = Fls_BlankCheckRxDataBuf;
-    uint8  max_len      = sizeof(Fls_BlankCheckRxDataBuf) / sizeof(Fls_BlankCheckRxDataBuf[0]);
+    uint32 max_len      = sizeof(Fls_BlankCheckRxDataBuf) / sizeof(Fls_BlankCheckRxDataBuf[0]);
 
     if ((Fls_DrvObj.spiHandle != NULL_PTR) && (len <= max_len))
     {
@@ -1490,7 +1490,7 @@ Std_ReturnType Fls_norCompare(uint32 actualChunkSize)
     uint32 len          = actualChunkSize;
     uint8 *readData_buf = Fls_CompareRxDataBuf;
     uint8 *expData_Buf  = Fls_DrvObj.ramAddr;
-    uint8  max_len      = sizeof(Fls_CompareRxDataBuf) / sizeof(Fls_CompareRxDataBuf[0]);
+    uint32 max_len      = sizeof(Fls_CompareRxDataBuf) / sizeof(Fls_CompareRxDataBuf[0]);
 
     if ((Fls_DrvObj.spiHandle != NULL_PTR) && (len <= max_len))
     {
